+++ src/imio/smartweb/common/caching.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from plone import api
+from urllib.parse import urlparse
 
 import logging
 import os
@@ -12,8 +13,10 @@ logger = logging.getLogger("imio.smartweb.common")
 def ban_physicalpath(request, physical_path):
     portal = api.portal.get()
     caching_servers = os.environ.get("CACHING_SERVERS", "").split(" ")
-    forwarded_host = request.get("X-Forwarded-Host", "") or request.get("HTTP_HOST", "")
-    headers = {"Host": forwarded_host}
+    website_hostname = (
+        os.environ.get("WEBSITE_HOSTNAME", "") or urlparse(portal.absolute_url()).netloc
+    )
+    headers = {"Host": website_hostname}
     len_portal_path = len(portal.getPhysicalPath())
     path = "/".join(physical_path[len_portal_path:])
     for caching_server in caching_servers:
@@ -24,6 +27,6 @@ def ban_physicalpath(request, physical_path):
         else:
             ban_url = f"http://{caching_server}"
         logger.info(
-            f"## BAN ## X-Forwarded-Host : {forwarded_host} ## ban_url : {ban_url}"
+            f"## BAN ## website_hostname : {website_hostname} ## ban_url : {ban_url}"
         )
         requests.request("BAN", ban_url, headers=headers)